Add the ability to search by name in the admin area
Since we have to preserve the case of the name we can't just lowercase
it like we do with the email so to achieve case-insensitiveness we
need to create an expression index on the name. Since support for these
indexes isn't available until Rails 5.0 we need to use SQL to generate
the index and override `index_exists?` because it doesn't appear in the
list of indexes returned by Active Record.

https://www.pivotaltracker.com/story/show/120103125